Folder restructuring
restructed test and include folder
