+++ test/CMakeLists.txt
@@ -23,7 +23,7 @@ if(CMAKE_COMPILER_IS_GNUCXX)
 endif()
 
 
-include_directories(../proto)
+include_directories(../uprotocol/proto)
 add_executable(json_serializer_test
 		${TESTS_PATH}/json/json_serializer_test.cpp
 		${CGREEN_INCLUDE_DIR}
