+++ uprotocol/cloudevent/serialize/binary_serializer.h
@@ -24,7 +24,7 @@
 #include <memory>
 #include <optional>
 
-#include "../model/cloud_event.h"
+#include "uprotocol/cloudevent/datamodel/cloud_event.h"
 #include "base64.h"
 #include "google/protobuf/util/time_util.h"
 #include "spdlog/spdlog.h"
