+++ uprotocol/tools/byte_order.cpp
