+++ uprotocol/tools/byte_order.h
