+++ uprotocol/tools/tools.h
