+++ uprotocol/uri/datamodel/uri_entity.h
