+++ uprotocol/uri/datamodel/uri_resource.h
