+++ uprotocol/uuid/random_gen.h
