+++ uprotocol/uuid/uuid_gen.cpp
