+++ uprotocol/uuid/uuid_gen.h
