+++ uprotocol/uuid/uuid_v8.h
